ai-learning-phase3/pyproject.toml
[project]
name = "ai-learning-phase3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "chromadb>=1.5.8",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.16",
    "langchain-text-splitters>=1.1.2",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
